Create t-sort
Performs a topological sort on a list of vertices. Takes a python list as with an even number of items, representing a graph, as input. (Items in list can be any python datatype.) If there are no cycles in the graph, the algorithm performs a sort. Outputs sorted items as string.